+++ src/AnnotatedElement.php
@@ -1,0 +1,14 @@
+<?php
+namespace Annotations;
+
+interface AnnotatedElement {
+	function getAnnotatedElementType();
+
+	function getAnnotations();
+
+	function getAnnotation($annotationClass);
+
+	function hasAnnotation($annotationClass);
+
+	function getDocComment();
+}
