+++ src/AnnotationParseException.php
@@ -1,0 +1,4 @@
+<?php
+namespace Annotations;
+
+class AnnotationParseException extends Exception {}
