+++ src/AnnotationParser.php
@@ -1,222 +1,5 @@
 <?php
-
-/*
-
-Copyright (c) 2013 by Matt Zabriskie
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-
-*/
-
-abstract class Annotation {
-	public $value;
-
-	function __construct($value = null) {
-		$this->value = $value;
-	}
-
-	function value() {
-		return $this->value;
-	}
-}
-
-/**
- * @AnnotationTarget(value=AnnotatedElementType::ANNOTATION_TYPE)
- */
-class AnnotationTarget extends Annotation {
-	function value() {
-		if ($this->value != null && is_array($this->value)) {
-			$bitmask = 0;
-			foreach ($this->value as $bit) {
-				$bitmask &= $bit;
-			}
-			$this->value = $bitmask;
-		}
-
-		return $this->value;
-	}
-}
-
-interface AnnotatedElementType {
-	/**
-	 * Annotation type declaration
-	 */
-	const ANNOTATION_TYPE = 1;
-
-	/**
-	 * Constructor declaration
-	 */
-	const CONSTRUCTOR = 2;
-
-	/**
-	 * Method declaration
-	 */
-	const METHOD = 4;
-
-	/**
-	 * Property declaration (includes constants)
-	 */
-	const PROPERTY = 8;
-
-	/**
-	 * Class (including annotation type), or interface declaration
-	 */
-	const TYPE = 16;
-}
-
-interface AnnotatedElement {
-	function getAnnotatedElementType();
-
-	function getAnnotations();
-
-	function getAnnotation($annotationClass);
-
-	function hasAnnotation($annotationClass);
-
-	function getDocComment();
-}
-
-class AnnotatedReflectionClass extends ReflectionClass implements AnnotatedElement {
-
-	function getAnnotatedElementType() {
-		return (AnnotatedElementType::TYPE | ($this->isSubClassOf('Annotation') ? AnnotatedElementType::ANNOTATION_TYPE : 0));
-	}
-
-	private $annotations;
-
-	function getAnnotations() {
-		if ($this->annotations == null) {
-			$this->annotations = AnnotationParser::parse($this);
-		}
-		return $this->annotations;
-	}
-
-	function getAnnotation($annotationClass) {
-		$all = $this->getAnnotations();
-		return (isset($all[$annotationClass]) ? $all[$annotationClass] : null);
-	}
-
-	function hasAnnotation($annotationClass) {
-		return $this->getAnnotation($annotationClass) != null;
-	}
-
-	private $methods;
-
-	function getMethods() {
-		if ($this->methods == null) {
-			$methods = parent::getMethods();
-			$this->methods = array();
-			foreach ($methods as $method) {
-				$this->methods[$method->getName()] = new AnnotatedReflectionMethod($this->getName(), $method->getName());
-			}
-		}
-		return $this->methods;
-	}
-
-	function getMethod($name) {
-		$all = $this->getMethods();
-		return (isset($all[$name]) ? $all[$name] : null);
-	}
-
-	function hasMethod($name) {
-		return $this->getMethod($name) != null;
-	}
-
-	private $properties;
-
-	function getProperties() {
-		if ($this->properties == null) {
-			$properties = parent::getProperties();
-			$this->properties = array();
-			foreach ($properties as $property) {
-				$this->properties[$property->getName()] = new AnnotatedReflectionProperty($this->getName(), $property->getName());
-			}
-		}
-		return $this->properties;
-	}
-
-	function getProperty($name) {
-		$all = $this->getProperties();
-		return (isset($all[$name]) ? $all[$name] : null);
-	}
-
-	function hasProperty($name) {
-		return $this->getProperty($name) != null;
-	}
-
-}
-
-class AnnotatedReflectionMethod extends ReflectionMethod implements AnnotatedElement {
-
-	function getAnnotatedElementType() {
-		return (AnnotatedElementType::METHOD | ($this->isConstructor() ? AnnotatedElementType::CONSTRUCTOR : 0));
-	}
-
-	private $annotations;
-
-	function getAnnotations() {
-		if ($this->annotations == null) {
-			$this->annotations = AnnotationParser::parse($this);
-		}
-		return $this->annotations;
-	}
-
-	function getAnnotation($annotationClass) {
-		$all = $this->getAnnotations();
-		return (isset($all[$annotationClass]) ? $all[$annotationClass] : null);
-	}
-
-	function hasAnnotation($annotationClass) {
-		return $this->getAnnotation($annotationClass) != null;
-	}
-
-}
-
-class AnnotatedReflectionProperty extends ReflectionProperty implements AnnotatedElement {
-
-	function getAnnotatedElementType() {
-		return AnnotatedElementType::PROPERTY;
-	}
-
-	private $annotations;
-
-	function getAnnotations() {
-		if ($this->annotations == null) {
-			$this->annotations = AnnotationParser::parse($this);
-		}
-		return $this->annotations;
-	}
-
-	function getAnnotation($annotationClass) {
-		$all = $this->getAnnotations();
-		return (isset($all[$annotationClass]) ? $all[$annotationClass] : null);
-	}
-
-	function hasAnnotation($annotationClass) {
-		return $this->getAnnotation($annotationClass) != null;
-	}
-
-}
-
-class AnnotationParseException extends Exception {}
-
-class AnnotationTargetException extends Exception {}
+namespace Annotations;
 
 class AnnotationParser {
 	// Tags used for generating PHP Docs (http://www.phpdoc.org/)
