+++ src/AnnotationTargetException.php
@@ -1,0 +1,4 @@
+<?php
+namespace Annotations;
+
+class AnnotationTargetException extends Exception {}
